Move caching logic into ApiHandler
Perform generic request caching within the ServeHTTP method of the API handler.

The new flow for a request is that when it comes in the APIHandler checks for
two methods on the appropriate server struct - the raw method name, and
CacheableMethodName. If the method is prefixed with Cacheable, it will
attempt to retrieve a value for the request from mecached and if none is found
will store the result of the subsequent call.

References #73